Fixing test case Denise/Sprites/blocked/blocked4 (OCS bug) #799

+++ Emulator/Components/Agnus/Sequencer/Sequencer.h
@@ -160,7 +160,7 @@ public:
     DDFState ddf;
 
     // Remembers the cycle when BPRUN goes up the first time
-    u16 bprunUp;
+    isize bprunUp;
 
 
     //
